PMM-12375 add version prop to ProxySQL service
PMM-12375 move infoRequest to serviceInfoBroker
