+++ api/inventorypb/json/client/services/add_proxy_sql_service_responses.go
@@ -449,6 +449,9 @@ type AddProxySQLServiceOKBodyProxysql struct {
 
 	// Custom user-assigned labels.
 	CustomLabels map[string]string `json:"custom_labels,omitempty"`
+
+	// ProxySQL version.
+	Version string `json:"version,omitempty"`
 }
 
 // Validate validates this add proxy SQL service OK body proxysql
