+++ api/inventorypb/json/client/services/get_service_responses.go
@@ -973,6 +973,9 @@ type GetServiceOKBodyProxysql struct {
 
 	// Custom user-assigned labels.
 	CustomLabels map[string]string `json:"custom_labels,omitempty"`
+
+	// ProxySQL version.
+	Version string `json:"version,omitempty"`
 }
 
 // Validate validates this get service OK body proxysql
