+++ api/inventorypb/json/client/services/list_services_responses.go
@@ -1102,6 +1102,9 @@ type ListServicesOKBodyProxysqlItems0 struct {
 
 	// Custom user-assigned labels.
 	CustomLabels map[string]string `json:"custom_labels,omitempty"`
+
+	// ProxySQL version.
+	Version string `json:"version,omitempty"`
 }
 
 // Validate validates this list services OK body proxysql items0
