+++ api/inventorypb/services.proto
@@ -133,6 +133,8 @@ message ProxySQLService {
   string replication_set = 8;
   // Custom user-assigned labels.
   map<string, string> custom_labels = 9;
+  // ProxySQL version.
+  string version = 11;
 }
 
 // HAProxyService represents a generic HAProxy service instance.
