+++ api/swagger/swagger-dev.json
@@ -16294,6 +16294,11 @@
                         "type": "string"
                       },
                       "x-order": 9
+                    },
+                    "version": {
+                      "description": "ProxySQL version.",
+                      "type": "string",
+                      "x-order": 10
                     }
                   },
                   "x-order": 0
@@ -16855,6 +16860,11 @@
                         "type": "string"
                       },
                       "x-order": 9
+                    },
+                    "version": {
+                      "description": "ProxySQL version.",
+                      "type": "string",
+                      "x-order": 10
                     }
                   },
                   "x-order": 3
@@ -17313,6 +17323,11 @@
                           "type": "string"
                         },
                         "x-order": 9
+                      },
+                      "version": {
+                        "description": "ProxySQL version.",
+                        "type": "string",
+                        "x-order": 10
                       }
                     }
                   },
@@ -28458,6 +28473,11 @@
                         "type": "string"
                       },
                       "x-order": 9
+                    },
+                    "version": {
+                      "description": "ProxySQL version.",
+                      "type": "string",
+                      "x-order": 10
                     }
                   },
                   "x-order": 0
